Create Assignment 6.Rmd
This assignment focused on using PUMS public use data to understand the impact of housing costs for renters in Sonoma County. This assignment also focused on using SF open data that describes parcel availability, parcel use, and parcel potential to understand current housing use in the Mission District of San Francisco. I do not own any of the course materials for CEE 218X. I do not own any of the data sources. This assignment is for educational purposes only.